backend/requirements.txt
# Framework Principal
django==4.2.11

# Conector para o Banco de Dados PostgreSQL
 psycopg2

# Servidor de Aplicação WSGI para Produção
gunicorn==21.2.0

# Framework para criar a API REST
djangorestframework==3.15.1

# Para permitir a comunicação entre Frontend e Backend
django-cors-headers==4.3.1

# Biblioteca para gerar os relatórios em PDF
reportlab==4.1.0

# Para gerenciar variáveis de ambiente (chaves secretas, etc.)
python-dotenv==1.0.1

dj-database-url==0.5.0
whitenoise==6.6.0
# Framework Principal
django==4.2.11

# ... (outras dependências permanecem iguais)
dj-database-url==0.5.0
whitenoise==6.6.0

# REMOVA a linha do jazzmin e ADICIONE estas duas:
django-admin-interface==0.26.1
django-colorfield==0.9.0
djangorestframework-simplejwt==5.3.1